+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Results are now sorted by severity in the short report.
 - The Credential Guard check now returns something only if the configuration is vulnerable.
 - The AlwaysInstallElevated check was refactored.
+- The Point and Print config check was refactored.
 
 ## 2023-08-14